Clean up and move datatypes
