+++ project/test/acceptance_tests.py
@@ -3,9 +3,8 @@
 import unittest
 import io
 import sys
-from project.data_types import IO
+from project.common.data_types import IO
 from project.runner.independent_runner import *
-from project.common.heater import *
 from project.test.utils.io_emulator import IOEmulator
 
 IOMapping = {
